+++ Communication/Packets/Incoming/Navigator/CanCreateRoomEvent.cs
@@ -1,13 +1,17 @@
 ﻿using Plus.Communication.Packets.Outgoing.Navigator;
 using Plus.HabboHotel.GameClients;
+using Plus.HabboHotel.Navigator;
 
 namespace Plus.Communication.Packets.Incoming.Navigator;
 
 internal class CanCreateRoomEvent : IPacketEvent
 {
-    public Task Parse(GameClient session, IIncomingPacket packet)
+    private readonly INavigatorService _navigatorService;
+
+    public CanCreateRoomEvent(INavigatorService navigatorService)
     {
-        session.Send(new CanCreateRoomComposer(false, 150));
-        return Task.CompletedTask;
+        _navigatorService = navigatorService;
     }
+
+    public Task Parse(GameClient session, IIncomingPacket packet) => _navigatorService.CanCreateRoom(session);
 }