+++ Communication/Packets/Incoming/Navigator/GetGuestRoomEvent.cs
@@ -1,19 +1,19 @@
 ﻿using Plus.Communication.Packets.Outgoing.Navigator;
 using Plus.HabboHotel.GameClients;
+using Plus.HabboHotel.Navigator;
 using Plus.HabboHotel.Rooms;
 
 namespace Plus.Communication.Packets.Incoming.Navigator;
 
 internal class GetGuestRoomEvent : IPacketEvent
 {
-    public Task Parse(GameClient session, IIncomingPacket packet)
+    private readonly INavigatorService _navigatorService;
+
+    public GetGuestRoomEvent(INavigatorService navigatorService)
     {
-        var roomId = packet.ReadUInt();
-        if (!RoomFactory.TryGetData(roomId, out var data))
-            return Task.CompletedTask;
-        var enter = packet.ReadInt() == 1;
-        var forward = packet.ReadInt() == 1;
-        session.Send(new GetGuestRoomResultComposer(session, data, enter, forward));
-        return Task.CompletedTask;
+        _navigatorService = navigatorService;
     }
+
+    public Task Parse(GameClient session, IIncomingPacket packet)
+        => _navigatorService.GetGuestRoom(session, packet.ReadUInt(), packet.ReadInt() == 1, packet.ReadInt() == 1);
 }